Shard PPR navigations tests into multiple files
Better for paralellization in CI.
